added correct docker-files also added many other files

+++ src/main/java/com/rob/FastQuestion/repo/QuestionFilesRepo.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface QuestionFilesRepo extends JpaRepository<QuestionFile, Long> {
-    List<QuestionFile> findByQuestionId(Integer id);
+    List<QuestionFile> findByQuestionId(Long id);
 }
